added checkstyle and nexus repos info to pom, bug fix in schedule test, moved tests to integration package
