+++ pom.xml
@@ -205,7 +205,30 @@
 						</goals>
 					</execution>
 				</executions>
-			</plugin>			
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-checkstyle-plugin</artifactId>
+				<version>${checkstyle.plugin.version}</version>
+				<executions>
+					<execution>
+						<id>validate</id>
+						<phase>validate</phase>
+						<configuration>
+							<configLocation>google_checks.xml</configLocation>
+							<consoleOutput>true</consoleOutput>
+							<violationSeverity>error</violationSeverity>
+							<skip>false</skip>
+							<linkXRef>false</linkXRef>
+							<includeTestSourceDirectory>true</includeTestSourceDirectory>
+							<outputFile>${project.build.directory}/edgex-checkstyles-result.xml</outputFile>
+						</configuration>
+						<goals>
+							<goal>check</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
 		</plugins>
 		<pluginManagement>
 			<plugins>
@@ -238,4 +261,11 @@
 			</plugins>
 		</pluginManagement>			
 	</build>
+	<repositories>
+		<repository>
+			<id>org.edgexfoundry</id>
+			<name>EdgeX Foundry Nexus Repository</name>
+			<url>${nexusproxy}/content/groups/public/</url>
+		</repository>
+	</repositories>
 </project>
