+++ src/test/java/org/edgexfoundry/controller/integration/DeviceClientTest.java
@@ -16,7 +16,7 @@
  * @version: 1.0.0
  *******************************************************************************/
 
-package org.edgexfoundry.controller;
+package org.edgexfoundry.controller.integration;
 
 import static org.edgexfoundry.test.data.DeviceData.TEST_LABELS;
 import static org.edgexfoundry.test.data.DeviceData.TEST_NAME;
@@ -31,6 +31,10 @@ import java.util.List;
 import javax.ws.rs.ClientErrorException;
 import javax.ws.rs.NotFoundException;
 
+import org.edgexfoundry.controller.AddressableClient;
+import org.edgexfoundry.controller.DeviceClient;
+import org.edgexfoundry.controller.DeviceProfileClient;
+import org.edgexfoundry.controller.DeviceServiceClient;
 import org.edgexfoundry.controller.impl.AddressableClientImpl;
 import org.edgexfoundry.controller.impl.DeviceClientImpl;
 import org.edgexfoundry.controller.impl.DeviceProfileClientImpl;
