+++ src/test/java/org/edgexfoundry/controller/integration/DeviceServiceClientTest.java
@@ -16,7 +16,7 @@
  * @version: 1.0.0
  *******************************************************************************/
 
-package org.edgexfoundry.controller;
+package org.edgexfoundry.controller.integration;
 
 import static org.edgexfoundry.test.data.ServiceData.TEST_LABELS;
 import static org.edgexfoundry.test.data.ServiceData.TEST_SERVICE_NAME;
@@ -33,6 +33,10 @@ import java.util.Set;
 import javax.ws.rs.ClientErrorException;
 import javax.ws.rs.NotFoundException;
 
+import org.edgexfoundry.controller.AddressableClient;
+import org.edgexfoundry.controller.DeviceClient;
+import org.edgexfoundry.controller.DeviceProfileClient;
+import org.edgexfoundry.controller.DeviceServiceClient;
 import org.edgexfoundry.controller.impl.AddressableClientImpl;
 import org.edgexfoundry.controller.impl.DeviceClientImpl;
 import org.edgexfoundry.controller.impl.DeviceProfileClientImpl;
